Switching accordion component to name, making accordion.tsx.old
