--- src/components/ui/tooltip.tsx
@@ -1,48 +1,0 @@
-"use client"
-
-import * as React from "react"
-import * as TooltipPrimitive from "@radix-ui/react-tooltip"
-
-import { cn } from "@/lib/utils"
-
-const TooltipProvider = TooltipPrimitive.Provider
-
-const Tooltip = TooltipPrimitive.Root
-
-function TooltipTrigger(
-  { className, ...props }: React.ComponentPropsWithoutRef<typeof TooltipPrimitive.Trigger>,
-  ref: React.Ref<React.ElementRef<typeof TooltipPrimitive.Trigger>>
-) {
-  return (
-    <TooltipPrimitive.Trigger
-      ref={ref}
-      className={cn(`${className ?? ""} text-inherit`)}
-      {...props}
-    />
-  );
-}
-
-const ForwardedTooltipTrigger = React.forwardRef(TooltipTrigger);
-
-function TooltipContent(
-  { className, sideOffset = 4, ...props }: React.ComponentPropsWithoutRef<typeof TooltipPrimitive.Content>,
-  ref: React.Ref<React.ElementRef<typeof TooltipPrimitive.Content>>
-) {
-  return (
-    <TooltipPrimitive.Content
-      ref={ref}
-      sideOffset={sideOffset}
-      className={cn(
-        "z-50 overflow-hidden rounded-md border bg-popover px-3 py-1.5 text-sm text-popover-foreground shadow-md animate-in fade-in-0 zoom-in-95 data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=closed]:zoom-out-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-const ForwardedTooltipContent = React.forwardRef(TooltipContent);
-ForwardedTooltipContent.displayName = TooltipPrimitive.Trigger.displayName;
-
-
-export { Tooltip, ForwardedTooltipTrigger as TooltipTrigger, ForwardedTooltipContent as TooltipContent, TooltipProvider }
